+++ sources/code/Plugins/AudioOpenAL/AudioClip.hpp
@@ -13,7 +13,7 @@ namespace Grindstone {
 				ALuint buffer,
 				std::uint32_t channelCount,
 				std::uint32_t sampleRate,
-				std::uint8_t bitsPerSample
+				std::uint16_t bitsPerSample
 			) : Asset(uuid, name),
 				buffer(buffer),
 				channelCount(channelCount),
@@ -23,7 +23,7 @@ namespace Grindstone {
 			ALuint buffer = -1;
 			std::uint32_t channelCount = 0;
 			std::uint32_t sampleRate = 0;
-			std::uint8_t bitsPerSample = 0;
+			std::uint16_t bitsPerSample = 0;
 
 			DEFINE_ASSET_TYPE("AudioClipAsset", AssetType::AudioClip)
 		};
